Relocated DFU request handlers
