+++ km_usb_test/descriptors.c
@@ -294,48 +294,6 @@ const __flash USB_StringDescriptor dfu_runtime_string = {
 	.bDescriptorType = USB_DTYPE_String,
 	.bString = u"Runtime"
 };
-
-void dfu_control_setup(void)
-{
-	switch (usb_setup.bRequest)
-	{
-		case DFU_DETACH:
-			dfu_runtime_cb_enter_dfu_mode();
-			return usb_ep0_out();
-
-		// read status
-		case DFU_GETSTATUS: {
-			uint8_t len = usb_setup.wLength;
-			if (len > sizeof(DFU_StatusResponse))
-				len = sizeof(DFU_StatusResponse);
-			DFU_StatusResponse *st = (DFU_StatusResponse *)ep0_buf_in;
-			st->bStatus = DFU_STATUS_OK;
-			st->bState = DFU_STATE_dfuIDLE;
-			st->bwPollTimeout[0] = 0;
-			st->bwPollTimeout[1] = 0;
-			st->bwPollTimeout[2] = 0;
-			st->iString = 0;
-			usb_ep0_in(len);
-			return usb_ep0_out();
-		}
-
-		// abort, clear status
-		case DFU_ABORT:
-		case DFU_CLRSTATUS:
-			usb_ep0_in(0);
-			return usb_ep0_out();
-
-		// read state
-		case DFU_GETSTATE:
-			ep0_buf_in[0] = 0;
-			usb_ep0_in(1);
-			return usb_ep0_out();
-
-		// unsupported requests
-		default:
-			return usb_ep0_stall();
-	}
-}
 #endif // USB_DFU_RUNTIME
