+++ km_usb_test/usb/dfu.h
@@ -86,11 +86,4 @@ enum {
 };
 
 
-extern void dfu_set_alternative(uint8_t alt);
-extern void dfu_control_setup(void);
-extern void dfu_control_out_completion(void);
-extern void dfu_control_in_completion(void);
-
-
-
 #endif /* DFU_H_ */
